Solve Day 2 with an arithmetic sum approach
This solution requires some pre-work on identifying number patterns for
creating an arithmetic series over the largest size of numbers in the
input. As a result, each number inbetween the start and end don't have to be
checked, improving things by a lot.

With this day too, created a new InputIterator - a generic one - for turning
the test data structs that I've been using into something that can be easily
iterated over with the familiar interface. It was a good learning exercise
in interfaces and more comptime tomfoolery.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>