Pair it all down to a single binary
This seemed simplier to have a unified runner and timer for all of the
results. And I can easily embed the input files into the binary, thus
eliminating the whole `Setup` class for common reader and whatnot.

I did change the Setup for an Answer type - to collect the answers for
reporting.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>